requirements.txt
Flask
Gunicorn
sqlalchemy
pymysql
python-dotenv
os